v0.1276 Translating to a dynamic loading environment
Added ExtensionLoader:
A class loader, that loads every class from a given package directory.
See JavaDoc

Added FileLoader:
Recursively loads files from a directory.

Added ObjectManager:
Should in the future handle all the objects, for persistence objects.

Added PersistenceFile:
When adding new classes for the object manager, this extension is
required as it's the super class for loadable objects.

Added TextureHandler:
Used in the future for 'cutting' pieces of an image of multiple use.