v0.13 Lighting and shadow system
Renamed class to LoadableEntityAnimation

Added Light:
Adds light to the stage. Can be added to GameObjects.

Modified GameWorld:
In the draw method:
FrameBufferObject is used to create realistic shadows as well as lights
through a lighting shader.

Modified GUITown:
Build menu is now stored in a window instead of a table.
Images instead of text.
Objects are split into categories.